+++ packages/client/src/links/httpLink.ts
@@ -1,5 +1,5 @@
 import { HttpLinkOptions, TRPCLink } from './core';
-import { httpRequest } from './internals/httpRequest';
+import { httpRequest } from '../internals/httpRequest';
 
 export function httpLink(opts: HttpLinkOptions): TRPCLink {
   const { url } = opts;
